Método Destroy() - Carrera (#35)
Se implementó el método destroy() de CarreraController y también se modificó el mismo método pero de UsuarioController.

+++ Backend/app/Http/Controllers/UsuarioController.php
@@ -10,14 +10,14 @@ use Image;
 
 class UsuarioController extends Controller
 {
-    
+    /*
     public function __construct()
     {
         $this->middleware(['permission:create user'], ['only' => ['create', 'store']]);
         $this->middleware(['permission:read user'], ['only' => 'index']);
         $this->middleware(['permission:update user'], ['only' => ['edit', 'update']]);
         $this->middleware(['permission:delete user'], ['only' => 'delete']);
-    }
+    }*/
 
     #Retorna listado de todos los usuarios
     public function index()
@@ -101,9 +101,9 @@ class UsuarioController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
-        $user->delete();
-        //dd($user);
-        //dd($user->delete_at);
-        return compact('user');
+        if ($user != null) {
+            $user->delete();
+            return compact('user');
+        }
     }
 }
